+++ src/web/assets/elementsform/js/paymentForm.js
@@ -2,7 +2,9 @@ class PaymentIntentsElements {
   constructor(publishableKey, container) {
     this.container = container;
     this.formNamespace = this.container.dataset.paymentFormNamespace;
-    this.stripeInstance = Stripe(publishableKey);
+    this.stripeInstance = Stripe(publishableKey, {
+      locale: this.container.dataset.locale
+    });
     this.elements = null;
     this.scenario = this.container.dataset.clientScenario;
     this.completePaymentActionUrl =